trap now uses curl to command qBittorrent shutdown
Requires the option: "Bypass authentication for clients on localhost" under the "Web UI" tab (as of version 4.6.4). Also "WebUI\Port=XXXX" must be in the qBittorrent.conf or port 8080 is used.

This also seemingly stops all data loss as long as all data is flushed/saved in ~15 seconds. After ~15 seconds SIGABRT is sent.